Updated push function to handle new python (3.12) environment and way SQLAlchemy version 2.0.31 handles things
Previous to that, the last function "update_records" would not return any error messages on server but would not push any of the data to the SQL db. It worked fine on personal dekstop or if using method 1 (brute approach) on server but that wipes the whole db. Method 2 (soft) is prefered but was not working. These changes now push correctly the data to it